// libs/core/forms/src/types.d.ts
export interface ValidationError {
    code: string;
    message: string;
    fieldPath?: string;
    resourceUuid?: string;
    suggestedFix?: string;
}